==> 1_intro.Rmd <==
# About the website

This is my Final assignment for the University of Helsinki course: [Introduction to Open Data Science (IODS).](https://courses.helsinki.fi/fi/hymy005/120776718) 

## Code repository

The _R-code_ used is stored and available for inspection in  [this GitHub repository](https://github.com/Juhous/IODS-final)

## Extra

Instead of following the instructions and making code-panel foldable, I decided to create my own [code-wrapper](https://github.com/Juhous/IODS-final/blob/master/showCodePanel.js), with some panels foldable and others always visible. If you find my solution too hard to evaluate, I'm happy to deliver the exercise as per your instructions.
