==> 2_eda.Rmd <==
# The data
```{R, echo = F}
#Hidden
source("helper_functions.R")
library(magrittr)
library(knitr)
library(kableExtra)
library(ggplot2)
library(GGally)
library(tidyr, warn.conflicts = F)
library(stringr)
library(dplyr, warn.conflicts = F)
#note
df <- read.csv("data/human.csv")
```
This dataset contains `r nrow(df)` obseravtions and `r ncol(df)-1` variables exploring dependancy between proxy variables of human developement and gender (in)equality by country. The data originates from http://hdr.undp.org/en/data. Details of data-wrangling can be found [here](https://github.com/Juhous/IODS-final/blob/master/create_human.R).

HDI attempts to describe quality of life -- or how developed a country is, and [comprises of](http://hdr.undp.org/en/content/human-development-index-hdi) life expectancy, education and GNI per capita. 

Female labour participation, education, reproductive health and parliamentary representation determine another developemental index, [Gender inequality index, (GII)](http://hdr.undp.org/en/content/gender-inequality-index-gii)

In this exercise, I explore whether determinats of GII  correlate with HDI. My hypothesis is that countries with low GII have better HDI. It'll be interesting to see which of the determinants of GII correlate most strongly with HDI. 

## Study variables

| Variable | Description |
|------|---------|
| HDI | Human developement index |
| GII | Gender inequality index |
| matMort | Maternal mortality ratio |
| adolBirthRate | Adolescent birth rate |
| reprParl | Percetange of female representatives in parliament |
| eduRatio | Ratio between edu2F and edu2M |
| labRatio | Ratio between labF and labM |

***

# Exploratory Data Analysis
## Numercial overview
```{r}
#Hidden
df %>% select(-country) %>% summaryKable() %>% 
  kable("html", align = "rrr", caption = "Study variable summary") %>%
  kable_styling(bootstrap_options = c("hover", "condensed"))
```

These values display no alarming features. Interestingly median of eduRatio is ~1, which I would've expected to be much smaller. 

## Graphical overview
```{R, out.width = "80%"}
#Hidden
df %>% gather(key="var", value = "value", HDI:labRatio) %>%
  mutate(var=ordered(var, names(df[2:8]))) %>%  #To retain order when facetting
  ggplot() +
  geom_density(aes(value)) + 
  facet_wrap(~var, scales="free") +
  labs(title = "Study variable distribution", 
       y = "Density", x="Value") +
  theme(axis.text.x = element_text(angle = 90))
```

Most variables are normally enough for my purpose. Maternal mortality si a bit skewed, and has long tail. Logarithmic conversion should make it more normal-like. 

```{R, fig.width = 6, fig.height = 2, out.width = "50%"}
#Hidden
# Save a plot of non-normally distributed variables
p1 <- df %>% 
  ggplot() +
  geom_density(aes(matMort)) + 
  labs(title = "Before log-conv", 
       y = "Density", x="Value") 

#Log conversion for non-normal values
df %<>% mutate(matMort = log10(matMort)) %>% 
  rename(matMort_log = matMort)

# Store a plot of new values
p2 <- df %>% 
  ggplot() +
  geom_density(aes(matMort_log)) + 
  labs(title = "After log-conv", 
       y = "Density", x="Value") 

#Show plots
multiplot(p1,p2, cols = 2)
```

Clearly the distribution is better after the log-conversion. 

## Overview of correlations
```{r, fig.width: 10, fig_height: 10}
#Hidden
p <- df %>% select(-country) %>% ggpairs(
  title = "Study variable overview",
  upper = list(continuous = wrap("cor", size = 4, color = "black")),
  lower = list(
    continuous = wrap("points", alpha = .2, size = .4),
    combo = wrap("facethist", bins = 20))) +
  theme(axis.text.x = element_text(
                  angle = 90,
                  color = "black",
                  size = 7,
                  vjust = .5),
      axis.text.y = element_text(color = "black", size = 7),
      strip.text = element_text(size = 8))

p[3,1] <-  p[3,1] + aes(color = "red")
p[4,1] <-  p[4,1] + aes(color = "red")
p[6,1] <-  p[6,1] + aes(color = "red")

p[1,3] <- p[1,3] + aes(fontface = 2)
p[1,4] <- p[1,4] + aes(fontface = 2)
p[1,6] <-  p[1,6] + aes(fontface = 2)

p
```

This graph contains a wealth of information and plenty of time should be used to familiarize oneself with it. From the scatterplots, we can conclude that many of the variables correlate with each other, and importantly, that no obvious non-linear relationships emerge from the data. 

Both, maternal mortality and adolescent birth rate, correlate strongly and negatively with HDI. 

EduRatio exhibits moderate positive correlation R=0,68. The relationship is not only linear; for most of HDI-spectrum there is no correlation, but before a certain threshold, a positive slope is visible (i.e. HDI correlates positively with eduRatio). 


The actual gender inequality index correlates moderately and negatively  with the HDI (and actually, all components of HDI, not shown). 

```{r, echo = F}
write.csv(df, "data/df.csv", row.names = F) # Need to access this dataframe on later .Rmds
```
